Finalized pre release add asset
