+++ example/lib/main.dart
@@ -8,7 +8,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(primarySwatch: Colors.blue),
+      theme: ThemeData.dark(),
       home: MyHomePage(),
     );
   }
@@ -54,7 +54,7 @@ class MyHomePage extends StatelessWidget {
               },
             ),
             Divider(),
-            Text('Custom Outline Button:'),
+            Text('Custom Outlined Button (Error):'),
             AsyncButtonBuilder(
               child: Text('Click Me'),
               loadingWidget: Text('Loading...'),
@@ -65,17 +65,17 @@ class MyHomePage extends StatelessWidget {
               },
               builder: (context, child, callback, buttonState) {
                 final buttonColor = buttonState.when(
-                  idling: () => Colors.pink,
-                  loading: () => Colors.yellow,
-                  completing: () => Colors.deepPurple,
-                  erroring: () => Colors.blueGrey,
+                  idling: () => Colors.yellow[200],
+                  loading: () => Colors.grey,
+                  completing: () => Colors.orangeAccent,
+                  erroring: () => Colors.orange,
                 );
 
                 return OutlinedButton(
                   child: child,
                   onPressed: callback,
                   style: OutlinedButton.styleFrom(
-                    primary: Colors.white,
+                    primary: Colors.black,
                     backgroundColor: buttonColor,
                   ),
                 );
@@ -84,30 +84,68 @@ class MyHomePage extends StatelessWidget {
             Divider(),
             Text('Custom Material Button:'),
             const SizedBox(height: 6.0),
-            Material(
-              color: Colors.lightBlue,
-              shape: StadiumBorder(),
-              child: AsyncButtonBuilder(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 16.0,
-                    vertical: 8.0,
-                  ),
-                  child: Text(
-                    'Click Me',
-                    style: TextStyle(color: Colors.white),
+            AsyncButtonBuilder(
+              child: Padding(
+                // Value keys are important as otherwise our custom transitions
+                // will have no way to differentiate between children.
+                key: ValueKey('lust'),
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 16.0,
+                  vertical: 8.0,
+                ),
+                child: Text(
+                  'Click Me',
+                  style: TextStyle(color: Colors.white),
+                ),
+              ),
+              loadingWidget: Padding(
+                key: ValueKey('envy'),
+                padding: const EdgeInsets.all(8.0),
+                child: SizedBox(
+                  height: 16.0,
+                  width: 16.0,
+                  child: CircularProgressIndicator(
+                    valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
                   ),
                 ),
-                onPressed: () async {
-                  await Future.delayed(Duration(seconds: 1));
-                },
-                builder: (context, child, callback, _) {
-                  return InkWell(
+              ),
+              completingWidget: Padding(
+                key: ValueKey('sloth'),
+                padding: const EdgeInsets.all(4.0),
+                child: Icon(
+                  Icons.check,
+                  color: Colors.purpleAccent,
+                ),
+              ),
+              onPressed: () async {
+                await Future.delayed(Duration(seconds: 2));
+              },
+              loadingSwitchInCurve: Curves.bounceInOut,
+              loadingTransitionBuilder: (child, animation) {
+                return SlideTransition(
+                  position: Tween<Offset>(
+                    begin: Offset(0, 1.0),
+                    end: Offset(0, 0),
+                  ).animate(animation),
+                  child: child,
+                );
+              },
+              builder: (context, child, callback, state) {
+                return Material(
+                  color: state.maybeWhen(
+                    completing: () => Colors.purple[100],
+                    orElse: () => Colors.blue,
+                  ),
+                  // This prevents the loading indicator showing below the
+                  // button
+                  clipBehavior: Clip.hardEdge,
+                  shape: StadiumBorder(),
+                  child: InkWell(
                     child: child,
                     onTap: callback,
-                  );
-                },
-              ),
+                  ),
+                );
+              },
             ),
             Divider(),
             Spacer(),
